Restore `Annotator`-decoding functions
