+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Serialisation/GoldenUtils.hs
@@ -23,6 +23,7 @@ import Cardano.Ledger.Binary (
   ToCBOR (..),
   Tokens (..),
   Version,
+  decodeFullAnnotator,
   decodeFullDecoder,
   decodeTerm,
   encCBOR,
@@ -38,7 +39,6 @@ import Data.String (fromString)
 import GHC.Stack
 import qualified Prettyprinter as Pretty
 import Test.Cardano.Ledger.Binary.TreeDiff (ansiDocToString, diffExpr)
-import Test.Cardano.Ledger.Shelley.Binary.Annotator (decodeFullAnnotator)
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (Assertion, assertFailure, testCase, (@?=))
