+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Serialisation/Golden/Encoding.hs
@@ -34,6 +34,7 @@ import Cardano.Ledger.BaseTypes (
   textToUrl,
  )
 import Cardano.Ledger.Binary (
+  Annotator,
   DecCBOR (..),
   DecCBORGroup (..),
   Decoder,
@@ -42,6 +43,7 @@ import Cardano.Ledger.Binary (
   EncCBORGroup (..),
   Tokens (..),
   byronProtVer,
+  decodeFullAnnotator,
   decodeFullDecoder,
   decodeMapTraverse,
   encCBOR,
@@ -152,7 +154,7 @@ import Test.Cardano.Ledger.Shelley.Serialisation.GoldenUtils (
   checkEncodingCBORAnnotated,
  )
 import Test.Cardano.Ledger.Shelley.Utils
-import Test.Cardano.Protocol.Binary.Annotator
+import Test.Cardano.Protocol.Binary.Annotator ()
 import Test.Cardano.Protocol.Crypto.VRF.Fake (WithResult (..))
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (assertFailure, testCase)
