+++ libs/cardano-ledger-api/cardano-ledger-api.cabal
@@ -94,7 +94,7 @@ library testlib
     base,
     bytestring,
     cardano-ledger-api,
-    cardano-ledger-binary:{cardano-ledger-binary, testlib},
+    cardano-ledger-binary,
     cardano-ledger-core:{cardano-ledger-core, testlib},
     containers,
