+++ libs/cardano-ledger-api/testlib/Test/Cardano/Ledger/Api/DebugTools.hs
@@ -14,6 +14,7 @@ import Cardano.Ledger.Binary (
   Version,
   decNoShareCBOR,
   decodeFull',
+  decodeFullAnnotator,
   decodeFullDecoder',
  )
 import Cardano.Ledger.Binary.Encoding (serialize')
@@ -22,7 +23,6 @@ import Control.Exception (Exception, throwIO)
 import Control.Monad.IO.Class (MonadIO (..))
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Lazy as LBS
-import Test.Cardano.Ledger.Binary.Annotator (decodeFullAnnotator)
 
 readCBORWith ::
   (MonadIO m, Exception e) => (Version -> BS.ByteString -> Either e a) -> Version -> FilePath -> m a
