+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary.hs
@@ -17,7 +17,6 @@ import qualified Data.ByteString.Lazy as BSL
 import Data.Proxy
 import qualified Data.Text as T
 import Data.Typeable
-import Test.Cardano.Ledger.Binary.Annotator (decodeFullAnnotator)
 import Test.Cardano.Ledger.Binary.RoundTrip (embedTripAnnExpectation)
 import Test.Hspec
 import Test.Hspec.QuickCheck (prop)
