+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/Annotator.hs
@@ -30,50 +30,7 @@ module Test.Cardano.Ledger.Binary.Annotator (
 import Cardano.Ledger.Binary
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Control.Monad.Except (Except, MonadError (throwError))
-import qualified Data.ByteString as BS
-import qualified Data.ByteString.Lazy as BSL
 import Data.Text (Text)
-
---------------------------------------------------------------------------------
--- Annotator
---------------------------------------------------------------------------------
-
--- | Same as `decodeFullDecoder`, except it provdes the means of passing portion or all
--- of the `BSL.ByteString` input argument to the decoding `Annotator`.
-decodeFullAnnotator ::
-  Version ->
-  Text ->
-  (forall s. Decoder s (Annotator a)) ->
-  BSL.ByteString ->
-  Either DecoderError a
-decodeFullAnnotator v lbl decoder bytes =
-  (\x -> runAnnotator x (Full bytes)) <$> decodeFullDecoder v lbl decoder bytes
-{-# INLINE decodeFullAnnotator #-}
-
--- | Same as `decodeFullDecoder`, decodes a Haskell value from a lazy
--- `BSL.ByteString`, requiring that the full ByteString is consumed, and
--- replaces `ByteSpan` annotations with the corresponding slice of the input as
--- a strict `BS.ByteString`.
-decodeFullAnnotatedBytes ::
-  Functor f =>
-  Version ->
-  Text ->
-  (forall s. Decoder s (f ByteSpan)) ->
-  BSL.ByteString ->
-  Either DecoderError (f BS.ByteString)
-decodeFullAnnotatedBytes v lbl decoder bytes =
-  annotationBytes bytes <$> decodeFullDecoder v lbl decoder bytes
-{-# INLINE decodeFullAnnotatedBytes #-}
-
-decodeFullAnnotatorFromHexText ::
-  Version ->
-  Text ->
-  (forall s. Decoder s (Annotator a)) ->
-  Text ->
-  Either DecoderError a
-decodeFullAnnotatorFromHexText v desc dec =
-  Plain.withHexText $ decodeFullAnnotator v desc dec . BSL.fromStrict
-{-# INLINE decodeFullAnnotatorFromHexText #-}
 
 -- | Translation function between values through a related binary representation. This
 -- function allows you to translate one type into another (or the same one) through their
