+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/Cddl.hs
@@ -46,7 +46,6 @@ import System.Process.Typed (
   setStdin,
  )
 import Test.Cardano.Ledger.Binary (decoderEquivalenceExpectation)
-import Test.Cardano.Ledger.Binary.Annotator (decodeFullAnnotator)
 import Test.Cardano.Ledger.Binary.RoundTrip
 import Test.Hspec
 import UnliftIO.Temporary (withTempFile)
