+++ libs/cardano-ledger-binary/testlib/Test/Cardano/Ledger/Binary/RoundTrip.hs
@@ -74,7 +74,6 @@ import Data.Functor
 import Data.Proxy
 import qualified Data.Text as T
 import Data.Typeable
-import Test.Cardano.Ledger.Binary.Annotator (decodeFullAnnotator)
 import Test.Cardano.Ledger.Binary.Plain.RoundTrip (
   showFailedTermsWithReSerialization,
   showMaybeDecoderError,
